@@ -5,7 +5,7 @@ description: Flutter Dogfooding App to showcase Video SDK.
 
 environment:
   sdk: '>=3.0.0 <4.0.0'
-  flutter: '>=3.10.0'
+  flutter: '>=3.16.0'
 
 dependencies:
   crypto: ^3.0.3
@@ -49,6 +49,8 @@ dependencies:
   webview_flutter: ^4.5.0
   webviewx_plus: ^0.5.0
   pin_code_fields: ^8.0.1
+  stop_watch_timer: ^3.1.1
+  sticky_headers: ^0.3.0+2
 
 dev_dependencies:
   build_runner: ^2.4.6
