@@ -72,26 +72,32 @@ class DefaultFirebaseOptions {
   // );
 
   static const FirebaseOptions android = FirebaseOptions(
-    apiKey:
-        'AIzaSyAAsFzCi9L-E9CuULP64MO_FVd8mo0J8S4', // From "api_key" -> "current_key"
-    appId:
-        '1:264371954957:android:436b2b784d5a5c56bbfc44', // From "client" -> "client_info" -> "mobilesdk_app_id"
-    messagingSenderId:
-        '264371954957', // From "project_info" -> "project_number"
-    projectId: 'propertyside-test', // From "project_info" -> "project_id"
-    storageBucket:
-        'propertyside-test.appspot.com', // From "project_info" -> "storage_bucket"
+    apiKey: 'AIzaSyAAsFzCi9L-E9CuULP64MO_FVd8mo0J8S4',
+    appId: '1:264371954957:android:436b2b784d5a5c56bbfc44',
+    messagingSenderId: '264371954957',
+    projectId: 'propertyside-test',
+    storageBucket: 'propertyside-test.appspot.com',
   );
 
+  // static const FirebaseOptions ios = FirebaseOptions(
+  //   apiKey: 'AIzaSyCvN-HAjjnGuJS1sV5-XkhZ0BYnkxXZdPs',
+  //   appId: '1:347024607410:ios:ffe113a4b22025cd8c21ab',
+  //   messagingSenderId: '347024607410',
+  //   projectId: 'stream-video-9b586',
+  //   storageBucket: 'stream-video-9b586.appspot.com',
+  //   iosClientId:
+  //       '347024607410-rdemfqlplsgrglpuc12itra4e4npo1p7.apps.googleusercontent.com',
+  //   iosBundleId: 'io.getstream.video.flutter.dogfooding',
+  // );
   static const FirebaseOptions ios = FirebaseOptions(
-    apiKey: 'AIzaSyCvN-HAjjnGuJS1sV5-XkhZ0BYnkxXZdPs',
-    appId: '1:347024607410:ios:ffe113a4b22025cd8c21ab',
-    messagingSenderId: '347024607410',
-    projectId: 'stream-video-9b586',
-    storageBucket: 'stream-video-9b586.appspot.com',
+    apiKey: 'AIzaSyDvP5JDD69_H9TBd_rCac0ypd-SFr3CtBk',
+    appId: '1:264371954957:ios:8d269be8a9538622bbfc44',
+    messagingSenderId: '264371954957',
+    projectId: 'propertyside-test',
+    storageBucket: 'propertyside-test.appspot.com',
     iosClientId:
-        '347024607410-rdemfqlplsgrglpuc12itra4e4npo1p7.apps.googleusercontent.com',
-    iosBundleId: 'io.getstream.video.flutter.dogfooding',
+        '264371954957-bbiqpnfb8f6566ojqp2gntj5t4vt5kef.apps.googleusercontent.com',
+    iosBundleId: 'com.propertyside.clientapp',
   );
 
   static const FirebaseOptions macos = FirebaseOptions(
