@@ -6,11 +6,21 @@ part of 'routes.dart';
 // GoRouterGenerator
 // **************************************************************************
 
-List<RouteBase> get $appRoutes =>
-    [$homeRoute, $loginRoute, $lobbyRoute, $callRoute, $channelListRoute];
+List<RouteBase> get $appRoutes => [
+      $homeRoute,
+      $channelListRoute,
+      $channelPageRoute,
+      $homePropertyRoute,
+      $profileRoute,
+      $teamRoute,
+      $loginRoute,
+      $oTPRoute,
+      $lobbyRoute,
+      $callRoute,
+    ];
 
 RouteBase get $homeRoute => GoRouteData.$route(
-      path: '/',
+      path: '/home',
       name: 'home',
       factory: $HomeRouteExtension._fromState,
     );
@@ -19,7 +29,7 @@ extension $HomeRouteExtension on HomeRoute {
   static HomeRoute _fromState(GoRouterState state) => HomeRoute();
 
   String get location => GoRouteData.$location(
-        '/',
+        '/home',
       );
 
   void go(BuildContext context) => context.go(location);
@@ -43,6 +53,36 @@ extension $ChannelListRouteExtension on ChannelListRoute {
 
   String get location => GoRouteData.$location(
         '/channelList',
+      );
+
+  void go(BuildContext context) => context.go(location);
+
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location);
+
+  void replace(BuildContext context) => context.replace(location);
+}
+
+RouteBase get $channelPageRoute => GoRouteData.$route(
+      path: '/channel',
+      name: 'channel',
+      factory: $ChannelPageRouteExtension._fromState,
+    );
+
+extension $ChannelPageRouteExtension on ChannelPageRoute {
+  static ChannelPageRoute _fromState(GoRouterState state) => ChannelPageRoute(
+        selectedMembers: state.uri.queryParametersAll['selected-members']!
+            .map((e) => e)
+            .toList(),
+      );
+
+  String get location => GoRouteData.$location(
+        '/channel',
+        queryParams: {
+          'selected-members': selectedMembers.map((e) => e).toList(),
+        },
       );
 
   void go(BuildContext context) => context.go(location);
@@ -79,8 +119,54 @@ extension $HomePropertyRouteExtension on HomePropertyRoute {
   void replace(BuildContext context) => context.replace(location);
 }
 
+RouteBase get $profileRoute => GoRouteData.$route(
+      path: '/profile',
+      name: 'profile',
+      factory: $ProfileRouteExtension._fromState,
+    );
+
+extension $ProfileRouteExtension on ProfileRoute {
+  static ProfileRoute _fromState(GoRouterState state) => ProfileRoute();
+
+  String get location => GoRouteData.$location(
+        '/profile',
+      );
+
+  void go(BuildContext context) => context.go(location);
+
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location);
+
+  void replace(BuildContext context) => context.replace(location);
+}
+
+RouteBase get $teamRoute => GoRouteData.$route(
+      path: '/team',
+      name: 'team',
+      factory: $TeamRouteExtension._fromState,
+    );
+
+extension $TeamRouteExtension on TeamRoute {
+  static TeamRoute _fromState(GoRouterState state) => TeamRoute();
+
+  String get location => GoRouteData.$location(
+        '/team',
+      );
+
+  void go(BuildContext context) => context.go(location);
+
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location);
+
+  void replace(BuildContext context) => context.replace(location);
+}
+
 RouteBase get $loginRoute => GoRouteData.$route(
-      path: '/login',
+      path: '/',
       name: 'login',
       factory: $LoginRouteExtension._fromState,
     );
@@ -89,7 +175,37 @@ extension $LoginRouteExtension on LoginRoute {
   static LoginRoute _fromState(GoRouterState state) => LoginRoute();
 
   String get location => GoRouteData.$location(
-        '/login',
+        '/',
+      );
+
+  void go(BuildContext context) => context.go(location);
+
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location);
+
+  void replace(BuildContext context) => context.replace(location);
+}
+
+RouteBase get $oTPRoute => GoRouteData.$route(
+      path: '/codeVerification',
+      name: 'codeVerification',
+      factory: $OTPRouteExtension._fromState,
+    );
+
+extension $OTPRouteExtension on OTPRoute {
+  static OTPRoute _fromState(GoRouterState state) => OTPRoute(
+        phoneNumber: state.uri.queryParameters['phone-number'],
+        verificationId: state.uri.queryParameters['verification-id'],
+      );
+
+  String get location => GoRouteData.$location(
+        '/codeVerification',
+        queryParams: {
+          if (phoneNumber != null) 'phone-number': phoneNumber,
+          if (verificationId != null) 'verification-id': verificationId,
+        },
       );
 
   void go(BuildContext context) => context.go(location);
